% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appManager.R
\name{app_isNew2}
\alias{app_isNew2}
\title{检验程序是否存在}
\usage{
app_isNew2(token = "36F0DB19-AC55-4062-B2DA-39DC39B297BE", app_name = "appTpl")
}
\arguments{
\item{token}{口令}

\item{app_name}{程序名}
}
\value{
返回值
}
\description{
检验程序是否存在
}
\examples{
app_isNew2()
}
