% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{dbConfig}
\alias{dbConfig}
\title{数据库选择器}
\usage{
dbConfig(FAppId = "vmdms", FType = "ERP", FRunEnv = "PRD")
}
\arguments{
\item{FAppId}{程序ID}

\item{FType}{类型}

\item{FRunEnv}{运行虞翊时尚发型}
}
\value{
返回值
}
\description{
数据库选择器
}
\examples{
dbConfig()
}
