% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{userDataRange_query}
\alias{userDataRange_query}
\title{查询用户的查看数据范围}
\usage{
userDataRange_query(
  token = "36F0DB19-AC55-4062-B2DA-39DC39B297BE",
  FappId = "jhdms",
  Fuser = "马斌"
)
}
\arguments{
\item{token}{口令}

\item{FappId}{程序}

\item{Fuser}{用户}
}
\value{
返回值
}
\description{
查询用户的查看数据范围
}
\examples{
userDataRange_query()
}
