% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{userGroup_formatter}
\alias{userGroup_formatter}
\title{用户格式化列表}
\usage{
userGroup_formatter(
  token = "36F0DB19-AC55-4062-B2DA-39DC39B297BE",
  FappId = "rdsdms",
  groups
)
}
\arguments{
\item{token}{口令}

\item{FappId}{程序}

\item{groups}{分组}
}
\value{
返回列表
}
\description{
用户格式化列表
}
\examples{
userGroup_formatter()
}
