% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{userPassword_reset}
\alias{userPassword_reset}
\title{重置用户密码}
\usage{
userPassword_reset(
  token = "36F0DB19-AC55-4062-B2DA-39DC39B297BE",
  FappId = "jhdms",
  Fuser = "马斌"
)
}
\arguments{
\item{token}{口令}

\item{FappId}{程序ID}

\item{Fuser}{用户名}
}
\value{
返回值
}
\description{
重置用户密码
}
\examples{
userPassword_reset()
}
