% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{userType_check}
\alias{userType_check}
\title{插入用户数据判断}
\usage{
userType_check(
  token = "36F0DB19-AC55-4062-B2DA-39DC39B297BE",
  FappId = "rdsdms",
  Fuser = "朱琼"
)
}
\arguments{
\item{token}{口令}

\item{FappId}{程序ID}

\item{Fuser}{用户}
}
\value{
返回列表
}
\description{
插入用户数据判断
}
\examples{
userType_check()
}
