% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appManager.R
\name{app_create2}
\alias{app_create2}
\title{创建程序}
\usage{
app_create2(
  app_name = "appTpl",
  token = "36F0DB19-AC55-4062-B2DA-39DC39B297BE"
)
}
\arguments{
\item{app_name}{程序名称}

\item{token}{口令}
}
\value{
返回值
}
\description{
创建程序
}
\examples{
app_create2()
}
